# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/products/store": {
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "description": "Create a new Product",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "brand",
                                    "description",
                                    "product_code",
                                    "category_id",
                                    "SKU",
                                    "min_stock_level",
                                    "price",
                                    "is_available",
                                    "tax_rate",
                                    "type_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sample Product"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Sample Brand"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Product Description"
                                    },
                                    "product_code": {
                                        "type": "string",
                                        "example": "PROD123"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "SKU": {
                                        "type": "string",
                                        "example": "SKU123"
                                    },
                                    "min_stock_level": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "99.99"
                                    },
                                    "is_available": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "tax_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "0.08"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IProduct"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/update/{id}": {
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Update a product",
                "description": "Update a Product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "brand",
                                    "description",
                                    "product_code",
                                    "category_id",
                                    "SKU",
                                    "min_stock_level",
                                    "price",
                                    "is_available",
                                    "tax_rate",
                                    "type_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sample Product"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Sample Brand"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Product Description"
                                    },
                                    "product_code": {
                                        "type": "string",
                                        "example": "PROD123"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "SKU": {
                                        "type": "string",
                                        "example": "SKU123"
                                    },
                                    "min_stock_level": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "99.99"
                                    },
                                    "is_available": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "tax_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "0.08"
                                    },
                                    "type_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IProduct"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/delete/{id}": {
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "description": "Delete a Product",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                },
                                "example": true
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "IProduct": {
                "required": [
                    "id",
                    "name",
                    "brand",
                    "description",
                    "product_code",
                    "category_id",
                    "SKU",
                    "min_stock_level",
                    "price",
                    "is_available",
                    "tax_rate",
                    "type_id",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Sample Product"
                    },
                    "brand": {
                        "type": "string",
                        "example": "Sample Brand"
                    },
                    "description": {
                        "type": "string",
                        "example": "Product Description"
                    },
                    "product_code": {
                        "type": "string",
                        "example": "PROD123"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "SKU": {
                        "type": "string",
                        "example": "SKU123"
                    },
                    "min_stock_level": {
                        "type": "integer",
                        "example": "10"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": "99.99"
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "tax_rate": {
                        "type": "number",
                        "format": "float",
                        "example": "0.08"
                    },
                    "type_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-13T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-13T14:30:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    }
}